+++ core/tests/bootstrap.php
@@ -228,7 +228,6 @@ namespace
 	require_once dirname(__DIR__) . '/contest.php';
 	require_once dirname(__DIR__) . '/album/type_registry.php';
 	require_once dirname(__DIR__) . '/policy/image_visibility.php';
-	require_once dirname(__DIR__) . '/event/legacy_contest_policy_listener.php';
 	require_once dirname(__DIR__) . '/file/file.php';
 	require_once dirname(__DIR__) . '/file/types/multiform.php';
 	require_once dirname(__DIR__) . '/album/album.php';